[SYCL] Fix a race condition when enqueueing an interop kernel
piKernelSetArg isn't supposed to be called from different threads
on the same kernel.
