+++ sycl/unittests/handler/SetArgForLocalAccessor.cpp
@@ -1,4 +1,5 @@
 #include <gtest/gtest.h>
+#include <helpers/KernelInteropCommon.hpp>
 #include <helpers/PiMock.hpp>
 
 #include <sycl/sycl.hpp>
@@ -9,135 +10,20 @@
 
 namespace {
 
-struct TestContext {
-  size_t localBufferArgSize = 0;
-
-  // SYCL RT has number of checks that all devices and contexts are consistent
-  // between kernel, kernel_bundle and other objects.
-  //
-  // To ensure that those checks pass, we intercept some PI calls to extract
-  // the exact PI handles of device and context used in queue creation to later
-  // return them when program/context/kernel info is requested.
-  pi_device deviceHandle;
-  pi_context contextHandle;
-
-  pi_program programHandle = createDummyHandle<pi_program>();
-
-  ~TestContext() { releaseDummyHandle<pi_program>(programHandle); }
-};
-
-TestContext GlobalContext;
-
-} // namespace
+size_t LocalBufferArgSize = 0;
 
 pi_result redefined_piKernelSetArg(pi_kernel kernel, pi_uint32 arg_index,
                                    size_t arg_size, const void *arg_value) {
-  GlobalContext.localBufferArgSize = arg_size;
-
-  return PI_SUCCESS;
-}
-
-pi_result after_piContextGetInfo(pi_context context, pi_context_info param_name,
-                                 size_t param_value_size, void *param_value,
-                                 size_t *param_value_size_ret) {
-  switch (param_name) {
-  case PI_CONTEXT_INFO_DEVICES:
-    if (param_value)
-      *static_cast<pi_device *>(param_value) = GlobalContext.deviceHandle;
-    if (param_value_size_ret)
-      *param_value_size_ret = sizeof(GlobalContext.deviceHandle);
-    break;
-  default:;
-  }
-
-  return PI_SUCCESS;
-}
-
-pi_result after_piProgramGetInfo(pi_program program, pi_program_info param_name,
-                                 size_t param_value_size, void *param_value,
-                                 size_t *param_value_size_ret) {
-
-  switch (param_name) {
-  case PI_PROGRAM_INFO_DEVICES:
-    if (param_value_size_ret)
-      *param_value_size_ret = sizeof(GlobalContext.deviceHandle);
-    if (param_value)
-      *static_cast<pi_device *>(param_value) = GlobalContext.deviceHandle;
-    break;
-  default:;
-  }
-
-  return PI_SUCCESS;
-}
-
-pi_result redefined_piProgramGetBuildInfo(pi_program program, pi_device device,
-                                          _pi_program_build_info param_name,
-                                          size_t param_value_size,
-                                          void *param_value,
-                                          size_t *param_value_size_ret) {
-  switch (param_name) {
-  case PI_PROGRAM_BUILD_INFO_BINARY_TYPE:
-    if (param_value_size_ret)
-      *param_value_size_ret = sizeof(pi_program_binary_type);
-    if (param_value)
-      *static_cast<pi_program_binary_type *>(param_value) =
-          PI_PROGRAM_BINARY_TYPE_EXECUTABLE;
-    break;
-  default:;
-  }
-
-  return PI_SUCCESS;
-}
-
-pi_result after_piContextCreate(const pi_context_properties *properties,
-                                pi_uint32 num_devices, const pi_device *devices,
-                                void (*pfn_notify)(const char *errinfo,
-                                                   const void *private_info,
-                                                   size_t cb, void *user_data),
-                                void *user_data, pi_context *ret_context) {
-  if (ret_context)
-    GlobalContext.contextHandle = *ret_context;
-  GlobalContext.deviceHandle = *devices;
-  return PI_SUCCESS;
-}
-
-pi_result after_piKernelGetInfo(pi_kernel kernel, pi_kernel_info param_name,
-                                size_t param_value_size, void *param_value,
-                                size_t *param_value_size_ret) {
-  switch (param_name) {
-  case PI_KERNEL_INFO_CONTEXT:
-    if (param_value_size_ret)
-      *param_value_size_ret = sizeof(GlobalContext.contextHandle);
-    if (param_value)
-      *static_cast<pi_context *>(param_value) = GlobalContext.contextHandle;
-    break;
-  case PI_KERNEL_INFO_PROGRAM:
-    if (param_value_size_ret)
-      *param_value_size_ret = sizeof(GlobalContext.programHandle);
-    if (param_value)
-      *(pi_program *)param_value = GlobalContext.programHandle;
-    break;
-  default:;
-  }
+  LocalBufferArgSize = arg_size;
 
   return PI_SUCCESS;
 }
 
 TEST(HandlerSetArg, LocalAccessor) {
   sycl::unittest::PiMock Mock;
-
+  redefineMockForKernelInterop(Mock);
   Mock.redefine<sycl::detail::PiApiKind::piKernelSetArg>(
       redefined_piKernelSetArg);
-  Mock.redefineAfter<sycl::detail::PiApiKind::piContextCreate>(
-      after_piContextCreate);
-  Mock.redefineAfter<sycl::detail::PiApiKind::piProgramGetInfo>(
-      after_piProgramGetInfo);
-  Mock.redefineAfter<sycl::detail::PiApiKind::piContextGetInfo>(
-      after_piContextGetInfo);
-  Mock.redefineAfter<sycl::detail::PiApiKind::piKernelGetInfo>(
-      after_piKernelGetInfo);
-  Mock.redefine<sycl::detail::PiApiKind::piProgramGetBuildInfo>(
-      redefined_piProgramGetBuildInfo);
 
   constexpr size_t Size = 128;
   sycl::queue Q;
@@ -154,5 +40,6 @@ TEST(HandlerSetArg, LocalAccessor) {
      CGH.single_task(Kernel);
    }).wait();
 
-  ASSERT_EQ(GlobalContext.localBufferArgSize, Size * sizeof(float));
+  ASSERT_EQ(LocalBufferArgSize, Size * sizeof(float));
 }
+} // namespace
