+++ sycl/unittests/thread_safety/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_sycl_unittest(ThreadSafetyTests OBJECT 
     HostAccessorDeadLock.cpp
+    InteropKernelEnqueue.cpp
 )
